Implemented (projectile) shooting functionality
I have it subtract a coin each time you click m1, I named the String action thing "click", this can be changed if too generic.  I tried to mess with collision to have it go away when it hits a wall but I wasn't getting it to work so I stopped.